remove mapp.layer changeend method
