+++ lib/layer/_layer.mjs
@@ -25,7 +25,6 @@ A mapp-layer object is a decorated JSON layer object which has been added to a m
 import decorate from './decorate.mjs';
 import formats from './format/_format.mjs';
 import fade from './fade.mjs';
-import changeEnd from './changeEnd.mjs'
 import featureHover from './featureHover.mjs';
 import * as featureFormats from './featureFormats.mjs';
 import * as featureFields from './featureFields.mjs';
@@ -38,7 +37,6 @@ import graduated from './themes/graduated.mjs';
 
 export default {
   decorate,
-  changeEnd,
   formats,
   featureFormats,
   featureFields,
