+++ lib/layer/decorate.mjs
@@ -30,7 +30,6 @@ The layer decorator method create mapp-layer typedef object from a json-layer.
 
 @returns {layer} Decorated Mapp Layer.
 */
-
 export default async function decorate(layer) {
 
   // Decorate layer format methods.
@@ -38,6 +37,8 @@ export default async function decorate(layer) {
 
   // If layer does not exist, return.
   if (!layer.L) return;
+
+  changeEndLoad(layer)
 
   // Assign show, hide, and other methods to the layer object.
   Object.assign(layer, {
@@ -102,9 +103,10 @@ export default async function decorate(layer) {
 }
 
 /**
-Shows the layer on the map.
 @function show
-@memberof module:/layer/decorate
+
+@description
+Shows the layer on the map.
 */
 function show() {
 
@@ -132,10 +134,11 @@ function show() {
 }
 
 /**
- * Hides the layer from the map.
- * @memberof module:layer/decorate
- * @function hide
- */
+@function hide
+
+@description
+Hides the layer from the map.
+*/
 function hide() {
 
   this.display = false;
@@ -156,11 +159,13 @@ function hide() {
 }
 
 /**
- * Returns the current table associated with the layer.
- * @memberof module:layer/decorate
- * @function tableCurrent
- * @returns {string} - The current table associated with the layer.
- */
+@function tableCurrent
+
+@description
+Returns the current table associated with the layer.
+
+@returns {string} The current table associated with the layer.
+*/
 function tableCurrent() {
 
   // Return the current table if it exists.
@@ -191,11 +196,13 @@ function tableCurrent() {
 }
 
 /**
- * Returns the current geometry associated with the layer.
- * @memberof module:layer/decorate
- * @function geomCurrent
- * @returns {string} - The current geometry associated with the layer.
- */
+@function geomCurrent
+
+@description
+Returns the current geometry associated with the layer.
+
+@returns {string} The current geometry associated with the layer.
+*/
 function geomCurrent() {
   // Return the current geometry if it exists.
   if (!this.geoms) return this.geom;
@@ -225,13 +232,15 @@ function geomCurrent() {
 }
 
 /**
- * Zooms to a specific extent on the map.
- * @memberof module:layer/decorate
- * @async
- * @function zoomToExtent
- * @param {Object} params - Parameters for zooming to extent.
- * @returns {Promise<boolean>} - A promise that resolves with a boolean indicating the success of the operation.
- */
+@function zoomToExtent
+@async
+
+@description
+Zooms to a specific extent on the map.
+
+@param {Object} params - Parameters for zooming to extent.
+@returns {Promise<boolean>} A promise that resolves with a boolean indicating the success of the operation.
+*/
 async function zoomToExtent(params) {
   // Zooms to a specific extent.
 
@@ -264,4 +273,35 @@ async function zoomToExtent(params) {
   );
 
   return true;
+}
+
+function changeEndLoad(layer) {
+
+  // The layer may be zoom level restricted.
+  layer.tables && layer.mapview.Map.getTargetElement().addEventListener('changeEnd', changeEnd)
+
+  function changeEnd() {
+
+    // Layer is out of zoom range.
+    if (!layer.tableCurrent()) {
+
+      if (layer.display) {
+
+        // Layer should be shown if possible.
+        layer.zoomDisplay = true
+        layer.hide()
+      }
+
+      return;
+    }
+
+    if (layer.zoomDisplay) {
+
+      // Prevents layer.show() being fired on zoom change within range.
+      delete layer.zoomDisplay
+
+      // Show layer if within zoomDisplay range.
+      layer.show()
+    }
+  }
 }
