+++ lib/layer/format/mvt.mjs
@@ -62,9 +62,6 @@ export default function MVT(layer) {
   // MVT query must not have a viewport, this is defined by the tile extent.
   delete layer.params.viewport
 
-  // Assigns changeEnd event/method for zoom restricted layers.
-  mapp.layer.changeEnd(layer)
-
   layer.reload = reload
 
   /**
