+++ lib/layer/format/vector.mjs
@@ -93,9 +93,6 @@ export default function vector(layer) {
     layer.L.setSource(layer.source)
   }
 
-  // Assigns changeEnd event/method for zoom restricted layers.
-  mapp.layer.changeEnd(layer)
-
   // Assign reload method to layer.
   layer.reload = reload;
